# docs/openapi.json
{
  "components": {
    "schemas": {
      "ContextPack": {
        "properties": {
          "adapter": {
            "default": "",
            "title": "Adapter",
            "type": "string"
          },
          "answerable": {
            "title": "Answerable",
            "type": "boolean"
          },
          "approved_page_count": {
            "default": 0,
            "title": "Approved Page Count",
            "type": "integer"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "title": "Evidence",
            "type": "array"
          },
          "graph": {
            "additionalProperties": {
              "items": {
                "additionalProperties": true,
                "type": "object"
              },
              "type": "array"
            },
            "title": "Graph",
            "type": "object"
          },
          "implementation": {
            "default": "",
            "title": "Implementation",
            "type": "string"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "orientation": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "title": "Orientation",
            "type": "array"
          },
          "page_count": {
            "default": 0,
            "title": "Page Count",
            "type": "integer"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "wiki_title": {
            "title": "Wiki Title",
            "type": "string"
          }
        },
        "required": [
          "query",
          "wiki_title",
          "answerable"
        ],
        "title": "ContextPack",
        "type": "object"
      },
...
      "GraphResponse": {
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "title": "GraphResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonRpcRequest": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "jsonrpc": {
            "default": "2.0",
            "title": "Jsonrpc",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          }
        },
        "required": [
          "method"
        ],
        "title": "JsonRpcRequest",
        "type": "object"
      },
      "JsonRpcResponse": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonRpcErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "jsonrpc": {
            "default": "2.0",
            "title": "Jsonrpc",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Result"
          }
        },
        "title": "JsonRpcResponse",
        "type": "object"
      },
...
      "QueryRequest": {
        "properties": {
          "include_drafts": {
            "default": false,
            "title": "Include Drafts",
            "type": "boolean"
          },
          "limit": {
            "default": 8,
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "query": {
            "default": "",
            "title": "Query",
            "type": "string"
          }
        },
        "title": "QueryRequest",
        "type": "object"
      },
...
      "WikiPage": {
        "properties": {
          "frontmatter": {
            "additionalProperties": true,
            "title": "Frontmatter",
            "type": "object"
          },
          "headings": {
            "items": {
              "type": "string"
            },
            "title": "Headings",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "links": {
            "items": {
              "type": "string"
            },
            "title": "Links",
            "type": "array"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "review_state": {
            "default": "unknown",
            "enum": [
              "approved",
              "reviewed",
              "verified",
              "draft",
              "proposed",
              "needs_review",
              "unknown"
            ],
            "title": "Review State",
            "type": "string"
          },
          "role": {
            "enum": [
              "hot",
              "index",
              "overview",
              "topic"
            ],
            "title": "Role",
            "type": "string"
          },
          "source_refs": {
            "items": {
              "type": "string"
            },
            "title": "Source Refs",
            "type": "array"
          },
          "status": {
            "default": "",
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "default": "",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "path",
          "role",
          "text"
        ],
        "title": "WikiPage",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Read-only HTTP, MCP-style JSON-RPC, MCP Streamable HTTP, and optional A2A-style message surface for LLMWiki Markdown folders.",
    "title": "LLMWiki Serve",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/graph": {
      "get": {
        "operationId": "graph_graph_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include_drafts",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Drafts",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Graph"
      }
    },
...
    "/mcp": {
      "post": {
        "operationId": "mcp_mcp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonRpcRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonRpcResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mcp"
      }
    },
    "/query": {
      "post": {
        "operationId": "query_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextPack"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query"
      }
    },
    "/read/{page_id}": {
      "get": {
        "operationId": "read_read__page_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_drafts",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Drafts",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/WikiPage"
                    },
                    {
                      "$ref": "#/components/schemas/ReadNotFoundResponse"
                    }
                  ],
                  "title": "Response Read Read  Page Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpDetailResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read"
      }
    },
    "/search": {
      "post": {
        "operationId": "search_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search"
      }
    },
    "/source-bundle": {
      "get": {
        "operationId": "source_bundle_source_bundle_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_drafts",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Drafts",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceBundleManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Source Bundle"
      }
    },
    "/source-refs": {
      "get": {
        "operationId": "source_refs_source_refs_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_drafts",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Drafts",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRefsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Source Refs"
      }
    }
  }
}